+++ CHANGELOG.md
@@ -38,6 +38,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - debian: preferred packages install and purge script
 - debian: 1 day system log rotation
 - debian: 1password install via package manager
+- debian: install latest software development tool binaries
 
 ### Changed
 - git: restructure multiple configs, using relative include file paths